requirements.txt
streamlit
langchain
langchain_groq
faiss-cpu
sentence-transformers
python-dotenv
PyPDF2
